refactor(tasks): add cli utils
