+++ tasks/dev.js
@@ -1,4 +1,4 @@
-import { isRunningFromCLI } from "../utils/task.js"
+import { isRunningFromCLI } from "../utils/cli.js"
 import { envFrom } from "../utils/env.js"
 import * as db from "./db.js"
 import * as vite from "./vite.js"
