+++ utils/cli.js
@@ -1,0 +1,8 @@
+import { resolve } from "path"
+import { fileURLToPath } from "url"
+
+
+
+export const isRunningFromCLI = 
+  (nodePath, modulePath) =>
+  resolve(nodePath) === resolve(fileURLToPath(modulePath))
